requirements.txt
pytest==7.4.0
assertpy == 1.1
allure-pytest==2.13.2
requests==2.31.0
